api-produtos/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
openai
cohere
tiktoken
requests
pandas
tqdm
